VolPred/requirements.txt
httpx
arch
joblib
pandas
numpy
python-dotenv
fastapi
uvicorn
apscheduler
sqlalchemy
